adds a lifetime model to handle api lifetime data
the api now delivers a data.lifetime object that
provides lifetime benefit values for each bar
visible in the graph. formerly we were overestimating
lifetime benefits because we were starting from a year,
and not a year plus a possible number of months based
on age or full-retirement year/month.