Fix migrations and standalone urls
The former migrations were created by South and imported that package.
Migrations have been squashed and coverted to native Django.
The former standalone url scheme didn't handle the /retirement/ url node
that exists on servers. This creates a separate set of urls just for
standalone and only uses them when STANDALONE is True.