fix tests that fail when run on certain days
One unit test began failing recently:

retirement_api.utils.tests.test_ss_utilities.UtilitiesTests.test_interpolate_for_past_fra

This seems to caused by the way that some of this project's tests make
use of datetime.today to determine the current day. This means that the
tests are running using slightly different data depending on the day on
which they are run. This makes it so that the edge cases of different
date calculations may work one day but not the next.

To fix this particular error, this change brings in the Python freezegun
package that lets you "freeze time" when running tests. This dependency
is only used when running tests, and won't be installed normally with
this app.

This change fixes this particular test, which relates to the
interpolation of benefits for retirees post-retirement age. I've added
two tests that result in 100% coverage of the method being tested, along
with some comments that attempt to explain the testing logic.