remove cf-icon fonts from repo and fix less path
This commit removes the copy of the Capital Framework icon font that was
in this repo. The gulp build task copies the cf-icon files under the
retirement_api's static/fonts path, which replaces these files anyway.

It also corrects how these files are specified in the Less/CSS. They
should be referred to under '../fonts', not '/static/retirement/icons'.
Use of a relative path is good here because there's no guarantee that
static assets will always be served under '/static/'.