change default webfont loading to use CDN
This commit modifies the default loading strategy for our licensed font
to use the Fonts.com CDN instead of looking for the file locally. This
means that when building and loading this repository locally, the
webfont will be loaded from the CDN. I've updated the README to indicate
how this behavior may be changed if you want to load the font locally.

This is necessary because making the default local loading means that
the built CSS includes references to a local webfonts directory that
will not exist in a built Python package. This then fails Django
collectstatic when run in a certain way.