Remove unused translation entry for "Search"
The Django translations file for this package includes one for the
English word "Search", translated to the Spanish word "búsqueda". This
translation is unused in this package, and conflicts with a different
translation in our parent consumerfinance.gov project ("Buscar"). [0]

This commit manually removes that translation from the Spanish django.po
file. I deliberately did not run Django makemessages because it appears
there are quite a few other translation entries that don't have code
equivalents, and doing so might result in a much larger set of changes.

All unit tests continue to pass, and English and Spanish versions of the
pages should still work properly.

[0] https://github.com/cfpb/consumerfinance.gov/blob/274b2522638ae7c2c381b632d48aa5969c54d019/cfgov/locale/es/LC_MESSAGES/django.po#L113-L114