Match the cfgov plurals config for translations
To fix a bug, Django 2.2.12 changed the way translation files are combined, and the
result changed how our retirement and cfgov Spanish translations worked together.

For the translations to work as before, both `django.po` files need to have the
same plural-translations scheme, which is defined at the top of the .po file.
cfgov had such a scheme defined (recently applied), and retirement did not,
putting them out of sync.

This adds a plural config line to retirement's django.po file so that the
definitions get properly merged.